start tracker earlier make bot reference directly available to command dispatcher
