+++ hangupsbot/hangupsbot.py
@@ -615,8 +615,23 @@ class HangupsBot(object):
 
         logging.debug("connected")
 
+        plugins.tracking.set_bot(self)
+        command.set_tracking(plugins.tracking)
+        command.set_bot(self)
+
+        plugins.tracking.start({ "module": "BASIC", "module.path": "commands.basic" })
+        import commands.basic
+        plugins.tracking.end()
+
+        plugins.tracking.start({ "module": "TAGS", "module.path": "commands.tagging" })
+        self.tags = tagging.tags(self)
+        import commands.tagging
+        plugins.tracking.end()
+
+        plugins.tracking.start({ "module": "HANDLERS", "module.path": "handlers" })
         self._handlers = handlers.EventHandler(self)
         handlers.handler.set_bot(self) # shim for handler decorator
+        plugins.tracking.end()
 
         self._user_list = yield from hangups.user.build_user_list(self._client,
                                                                   initial_data)
@@ -626,8 +641,9 @@ class HangupsBot(object):
                                                    self._user_list,
                                                    initial_data.sync_timestamp)
 
+        plugins.tracking.start({ "module": "PERMAMEM", "module.path": "permamem" })
         self.conversations = yield from permamem.initialise_permanent_memory(self)
-        self.tags = tagging.tags(self)
+        plugins.tracking.end()
 
         plugins.load(self, command)
